refactor(server): unify public API and eliminate deep imports
Refactored server module to use public API imports from other modules
and cleaned up __init__.py to expose only public symbols.

Changes:
- app.py: Replaced 7 deep imports with public API imports
  - Clients, config, models, observability, orchestra, runtime imports
  - Fixed inline imports: register_event_handlers, FailedGate (2x)
  - get_manager_usernames
  - Kept 2 exceptions: _resolve_manager_token (private), ServeStatusService (not exported)

- registry.py: Replaced 9 deep imports with public API imports
  - Clients, domain, environment, execution, models, orchestra, runtime imports
  - Fixed inline import: CodeagentBackend
  - Kept 5 exceptions: OrchestrationFacade, resolve_orchestra_repo_root,
    create_global_dispatch_coordinator, FailedGateProtocol, OrchestraSnapshot

- mcp.py: Replaced OrchestraStatusService import with public API import
  - Kept OrchestraSnapshot deep import (not exported from services)

- __init__.py: Cleaned up exports
  - Changed TYPE_CHECKING imports to use public API (vibe3.runtime instead of .orchestra_instance)
  - Updated _LAZY_IMPORTS to use public module paths
  - Removed 13 private symbols from __all__: _serialize_snapshot, _build_server,
    _build_server_with_launch_cwd, _build_async_serve_command, _start_async_serve,
    _setup_tailscale_webhook, _orchestra_tmux_session_exists, _kill_orchestra_tmux_session,
    _resolve_dispatcher_models_root, _resolve_orchestra_log_dir,
    _resolve_async_cli_override_root, _validate_pid_file, ORCHESTRA_TMUX_SESSION
  - Final __all__: 11 public symbols only

Test updates:
- test_mcp_server.py: Updated imports to use private symbols from actual modules
- test_serve.py: Updated FailedGate mock patches to patch vibe3.domain.FailedGate

Verification:
- All 60 targeted tests pass
- mypy: no issues found in 5 source files
- ruff: All checks passed

Result: 24 violations fixed, 7 deep imports remain as documented exceptions
(symbols not exported from target modules - out of scope).

Closes #2137
