+++ src/vibe3/server/mcp.py
@@ -8,10 +8,8 @@ from loguru import logger
 if TYPE_CHECKING:
     from mcp.server.fastmcp import FastMCP
 
-    from vibe3.services.orchestra_status_service import (
-        OrchestraSnapshot,
-        OrchestraStatusService,
-    )
+    from vibe3.services import OrchestraStatusService
+    from vibe3.services.orchestra_status_service import OrchestraSnapshot
 
 
 def _serialize_snapshot(snapshot: "OrchestraSnapshot") -> dict:
